added auth protections to actions in /server/actions/delete/actions.ts

+++ server/actions/delete/actions.ts
@@ -5,11 +5,16 @@ import {
   deleteFundTypeById,
 } from "@/server/supabase/functions/delete";
 import { revalidatePath } from "next/cache";
+import { auth } from "@clerk/nextjs/server";
 
 export async function DeleteFund(fundId: number, requestId: number) {
+  const { userId: clerkuserId } = auth();
+  if (!clerkuserId) {
+    throw new Error("User not authenticated");
+  }
   try {
     await deleteFundById(fundId);
-    await revalidatePath(`/admin/request/${requestId}/page`);
+    revalidatePath(`/admin/request/${requestId}/page`);
   } catch (error) {
     console.error(`Failed to delete fund with id ${fundId}:`, error);
     throw error;
@@ -17,9 +22,13 @@ export async function DeleteFund(fundId: number, requestId: number) {
 }
 
 export async function DeleteAgency(agencyId: number) {
+  const { userId: clerkuserId } = auth();
+  if (!clerkuserId) {
+    throw new Error("User not authenticated");
+  }
   try {
     await deleteAgencyById(agencyId);
-    await revalidatePath(`/admin/settings`);
+    revalidatePath(`/admin/settings`);
   } catch (error) {
     console.error(`Failed to delete agency with id ${agencyId}:`, error);
     throw error;
@@ -27,9 +36,13 @@ export async function DeleteAgency(agencyId: number) {
 }
 
 export async function DeleteFundType(fundTypeId: number) {
+  const { userId: clerkuserId } = auth();
+  if (!clerkuserId) {
+    throw new Error("User not authenticated");
+  }
   try {
     await deleteFundTypeById(fundTypeId);
-    await revalidatePath(`/admin/settings`);
+    revalidatePath(`/admin/settings`);
   } catch (error) {
     console.error(`Failed to delete fund type with id ${fundTypeId}:`, error);
     throw error;
